+++ src/features/landing/components/ApplySection.tsx
@@ -8,6 +8,41 @@ import SubTitle from './SubTitle';
 
 const subTitles = ['간단한 정보만 입력해도', '알바 지원이 가능합니다'];
 
+const floatingAnimations = [
+  {
+    y: [-10, -20, -10],
+    x: [-5, 10, -5],
+    scale: [1, 1.05, 1],
+    rotate: [-2, 3, -2],
+    duration: 3,
+    delay: 0,
+  },
+  {
+    y: [-15, -25, -15],
+    x: [8, -12, 8],
+    scale: [1, 1.08, 1],
+    rotate: [3, -4, 3],
+    duration: 2.5,
+    delay: 0.5,
+  },
+  {
+    y: [-8, -18, -8],
+    x: [-12, 15, -12],
+    scale: [1, 1.03, 1],
+    rotate: [-1, 2, -1],
+    duration: 3.5,
+    delay: 1,
+  },
+  {
+    y: [-12, -22, -12],
+    x: [15, -8, 15],
+    scale: [1, 1.06, 1],
+    rotate: [2, -3, 2],
+    duration: 2.8,
+    delay: 1.5,
+  },
+];
+
 const ApplySection = () => {
   return (
     <section className="relative z-10 flex h-screen flex-col justify-center overflow-hidden bg-transparent">
@@ -32,9 +67,30 @@ const ApplySection = () => {
           <SubTitle className="text-[#FFE1B2]" subTitles={subTitles} />
         </div>
         <div className="relative mt-126 flex flex-col gap-2">
-          <ul className="h-140 w-full">
-            {Array.from({ length: 4 }, (_, idx) => (
-              <li key={idx}>
+          <ul className="flex h-140 w-full justify-between px-50">
+            {floatingAnimations.map((animation, idx) => (
+              <motion.li
+                key={crypto.randomUUID()}
+                animate={{
+                  y: animation.y,
+                  x: animation.x,
+                  scale: animation.scale,
+                  rotate: animation.rotate,
+                }}
+                initial={{
+                  y: animation.y[0],
+                  x: animation.x[0],
+                  scale: animation.scale[0],
+                  rotate: animation.rotate[0],
+                }}
+                transition={{
+                  duration: animation.duration,
+                  repeat: Infinity,
+                  repeatType: 'reverse',
+                  ease: 'easeInOut',
+                  delay: animation.delay,
+                }}
+              >
                 <Image
                   alt={`apply-${idx + 1}`}
                   height={70}
@@ -42,10 +98,9 @@ const ApplySection = () => {
                   quality={90}
                   sizes="70px"
                   src={`/images/landing/apply-icon0${idx + 1}.png`}
-                  style={{ width: '70px', height: '70px' }}
                   width={70}
                 />
-              </li>
+              </motion.li>
             ))}
           </ul>
           <motion.div
